Move chart generation scripts into own package
