+++ charting_service/charts/generate_calories_output.py
