+++ charting_service/charts/generate_hoursofwork_output.py
