+++ charting_service/charts/generate_spendings_output.py
